Split tests into seperate files
In anticipation for the growth in number of tests
